REFACTOR: Removed get all pokemons from store
