+++ src/store/mutations/other.js
@@ -1,9 +1,9 @@
 import other from '@/store/state/other';
-import store from '@/lib/store';
 import {
   isDarkModeEnabled as isDarkModeEnabledInLS,
   toggleDarkMode as toggleDarkModeInLS,
 } from '@/lib/localStorage';
+import { getAll as getAllPokemons } from '@/store/mutations/pokemons';
 import { getAll as getAllGenerations } from '@/store/mutations/generations';
 import { getAll as getAllShapes } from '@/store/mutations/shapes';
 import { getAll as getAllColors } from '@/store/mutations/colors';
@@ -17,7 +17,7 @@ export async function initializeStore() {
       other.toggleIsDarkModeEnabled();
     }
     await Promise.all([
-      store.getAllPokemons(),
+      getAllPokemons(),
       getAllTypes(),
       getAllColors(),
       getAllShapes(),
